Fix menu keyboard handling, highscore input, and credits speeds
This fix re-enables saving all configuration options. The location of the
configuration file is always ~/.urban/ctrl.dat.
This fix also re-enables highscore input. Like in the original MS-DOS version
the player name is "Unknown" by default. The location of the highscore
file is always ~/.urban/hs.dat.
This fix also adds a delay to the credits scrolling loop, as scrolling speed
was dependent on the system speed in previous versions.